renamed sub template directory
