+++ source/04-templates/homepage/homepage.stories.jsx
@@ -6,7 +6,7 @@ import { withGlobalWrapper } from '../../../.storybook/decorators';
 import twigTemplate from './homepage.twig';
 import globalData from '../../00-config/storybook.global-data.yml';
 import ContentPlaceholder from '../../01-global/content-placeholder/content-placeholder';
-import { MessagesandTabs } from '../template-sections/messages-and-tabs/messages-and-tabs.stories.jsx';
+import { MessagesandTabs } from '../template-parts/messages-and-tabs/messages-and-tabs.stories.jsx';
 import { HeroImage } from '../../01-global/images/hero-image.stories.jsx';
 // Importing components to ensure their assets get loaded in Storybook when they
 // get referenced since Drupal loads them as a library.
