+++ source/04-templates/landing-page/landing-page.stories.jsx
@@ -6,7 +6,7 @@ import { withGlobalWrapper } from '../../../.storybook/decorators';
 import twigTemplate from './landing-page.twig';
 import globalData from '../../00-config/storybook.global-data.yml';
 import ContentPlaceholder from '../../01-global/content-placeholder/content-placeholder';
-import { MessagesandTabs } from '../template-sections/messages-and-tabs/messages-and-tabs.stories.jsx';
+import { MessagesandTabs } from '../template-parts/messages-and-tabs/messages-and-tabs.stories.jsx';
 
 const settings = {
   title: 'Templates/Landing Page',
