+++ source/04-templates/landing-page/landing-page.twig
@@ -1,4 +1,4 @@
-{% include '@templates/template-sections/messages-and-tabs/messages-and-tabs.twig' with {
+{% include '@templates/template-parts/messages-and-tabs/messages-and-tabs.twig' with {
   'is_published': is_published,
   'show_admin_info': show_admin_info,
   'admin_info': admin_info,
